Works but unclean
# Conflicts:
#	tests/datasets/dataframe/backends/xml/components/test_read.py
